Move common and utils to a core module
