+++ aces/idt/__init__.py
@@ -1,4 +1,4 @@
-from .common import (
+from aces.idt.core.common import (
     clf_processing_elements,
     error_delta_E,
     generate_reference_colour_checker,
@@ -6,14 +6,15 @@ from .common import (
     optimisation_factory_Oklab,
     png_compare_colour_checkers,
 )
-from .prosumer_camera import IDTGeneratorProsumerCamera
-from .utilities import (
+from aces.idt.core.utilities import (
     hash_file,
     list_sub_directories,
     mask_outliers,
     slugify,
     working_directory,
 )
+
+from .prosumer_camera import IDTGeneratorProsumerCamera
 
 __all__ = [
     "error_delta_E",
